// frontend/src/app/main/feature/netbase/infrastructure/models/CustomerModel.ts
export class CustomerModel {
    name: string;
    lastName: string;
    birthDate: string;
    documentNumber: string;
    phoneNumber: number;
    email: string;
    gender: string;
    city: string;
    hasTraveledAbroad: string;
    hasSpecificTravelDate: string;
    hasSpecificDestination: string;
    customerType: string;
    interestedInTicketsOrPackages: string;
    paymentMethod: string;
    createdAt: string;
    sourceUrl: string;
    attendedEvent: string;
    updatedAt?: string;
}